Convert dfp-api.spec.js to .ts
